fix: correct typo in badgemagic_module directory name
