+++ lib/badgemagic_module/bluetooth/connect_state.dart
@@ -1,5 +1,5 @@
-import 'package:badgemagic/bademagic_module/bluetooth/datagenerator.dart';
-import 'package:badgemagic/bademagic_module/bluetooth/write_state.dart';
+import 'package:badgemagic/badgemagic_module/bluetooth/datagenerator.dart';
+import 'package:badgemagic/badgemagic_module/bluetooth/write_state.dart';
 import 'package:get_it/get_it.dart';
 import '../../services/localization_service.dart';
 import '../../view/widgets/ble_progress_dialog.dart';
