+++ lib/badgemagic_module/bluetooth/scan_state.dart
@@ -1,6 +1,6 @@
 import 'dart:async';
-import 'package:badgemagic/bademagic_module/bluetooth/connect_state.dart';
-import 'package:badgemagic/bademagic_module/bluetooth/datagenerator.dart';
+import 'package:badgemagic/badgemagic_module/bluetooth/connect_state.dart';
+import 'package:badgemagic/badgemagic_module/bluetooth/datagenerator.dart';
 import 'package:badgemagic/providers/BadgeScanProvider.dart';
 import 'package:get_it/get_it.dart';
 import '../../services/localization_service.dart';
