+++ lib/badgemagic_module/models/data.dart
