+++ lib/badgemagic_module/models/mode.dart
