+++ lib/badgemagic_module/models/speed.dart
