+++ lib/badgemagic_module/utils/badge_text_storage.dart
@@ -1,7 +1,7 @@
 import 'dart:convert';
 import 'dart:io';
 
-import 'package:badgemagic/bademagic_module/utils/byte_array_utils.dart';
+import 'package:badgemagic/badgemagic_module/utils/byte_array_utils.dart';
 import 'package:path_provider/path_provider.dart';
 
 /// A utility class to store and retrieve the original text of badges
