+++ lib/badgemagic_module/utils/badge_utils.dart
