+++ lib/badgemagic_module/utils/image_utils.dart
@@ -1,6 +1,6 @@
 import 'dart:ui' as ui;
 
-import 'package:badgemagic/bademagic_module/utils/converters.dart';
+import 'package:badgemagic/badgemagic_module/utils/converters.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_svg/flutter_svg.dart';
