+++ lib/badgemagic_module/utils/toast_utils.dart
