+++ lib/view/qr_scan_screen.dart
@@ -1,5 +1,5 @@
-import 'package:badgemagic/bademagic_module/utils/qr_code_helper.dart';
-import 'package:badgemagic/bademagic_module/utils/toast_utils.dart';
+import 'package:badgemagic/badgemagic_module/utils/qr_code_helper.dart';
+import 'package:badgemagic/badgemagic_module/utils/toast_utils.dart';
 import 'package:badgemagic/l10n/app_localizations.dart';
 import 'package:badgemagic/services/localization_service.dart';
 import 'package:file_picker/file_picker.dart';
